+++ resources/js/components/frontend/nav/badge/favorites.tsx
